fix(storage-browser): move LOCATION_DETAIL_VIEW_HEADERS to constants file
