+++ packages/react-storage/src/components/StorageBrowser/views/LocationDetailView/getLocationDetailViewTableData/getFileRowContent.ts
@@ -1,12 +1,15 @@
+import { humanFileSize } from '@aws-amplify/ui';
+
 import { DataTableProps } from '../../../composables/DataTable';
 import { LocationData } from '../../../actions';
-import { LOCATION_DETAIL_VIEW_HEADERS } from './getLocationDetailViewTableData';
-import { humanFileSize } from '@aws-amplify/ui';
 import { displayText } from '../../../displayText/en';
+
+import { LOCATION_DETAIL_VIEW_HEADERS } from './constants';
 
 export const getFileRowContent = ({
   currentLocation,
   currentPath,
+
   isSelected,
   lastModified,
   rowId,
@@ -17,6 +20,7 @@ export const getFileRowContent = ({
 }: {
   currentLocation?: LocationData;
   currentPath: string;
+
   isSelected: boolean;
   lastModified: Date;
   rowId: string;
